directory() added, refactored a bit

+++ include/FileSystem.hpp
@@ -34,10 +34,10 @@ public:
     Status destroy(const std::string& file_name);
     std::pair<Status, size_t> open(const std::string& file_name);
     Status close(size_t index);
-    std::pair<Status, size_t> read(size_t index, char* mem_area, size_t count);
-    std::pair<Status, size_t> write(size_t index, char* mem_area, size_t count);
+    std::pair<Status, int> read(size_t index, char* mem_area, int count);
+    std::pair<Status, int> write(size_t index, char* mem_area, int count);
     Status lseek(size_t index, size_t pos);
-    std::unordered_map<std::string, size_t> directory();
+    std::unordered_map<std::string, int8_t> directory();
 
 private:
     IOSystem& iosystem;
